Extract AttachedEntity* bits from EntityAdapter into top-level. Add AttachedEntityHelper. Expose Document directly. Rename EntityAdapter.setMetadata() to EntityAdapter.attachMetadata().
