+++ components/nexus-orient/src/main/java/org/sonatype/nexus/orient/entity/EntityAdapter.java
@@ -12,13 +12,10 @@
  */
 package org.sonatype.nexus.orient.entity;
 
-import javax.annotation.Nonnull;
 import javax.inject.Inject;
 
 import org.sonatype.nexus.common.entity.Entity;
 import org.sonatype.nexus.common.entity.EntityId;
-import org.sonatype.nexus.common.entity.EntityMetadata;
-import org.sonatype.nexus.common.entity.EntityVersion;
 import org.sonatype.nexus.orient.RecordIdObfuscator;
 import org.sonatype.sisu.goodies.common.ComponentSupport;
 
@@ -28,12 +25,10 @@ import com.orientechnologies.orient.core.id.ORID;
 import com.orientechnologies.orient.core.metadata.schema.OClass;
 import com.orientechnologies.orient.core.metadata.schema.OSchema;
 import com.orientechnologies.orient.core.record.impl.ODocument;
-import com.orientechnologies.orient.core.version.ORecordVersion;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
 import static org.sonatype.nexus.common.entity.EntityHelper.id;
-import static org.sonatype.nexus.common.entity.EntityHelper.metadata;
 
 /**
  * Support for entity-adapter implementations.
@@ -138,7 +133,7 @@ public abstract class EntityAdapter<T extends Entity>
     catch (Exception e) {
       throw Throwables.propagate(e);
     }
-    setMetadata(entity, document);
+    attachMetadata(entity, document);
 
     return entity;
   }
@@ -158,7 +153,7 @@ public abstract class EntityAdapter<T extends Entity>
     catch (Exception e) {
       throw Throwables.propagate(e);
     }
-    setMetadata(entity, document);
+    attachMetadata(entity, document);
 
     return document.save();
   }
@@ -211,102 +206,10 @@ public abstract class EntityAdapter<T extends Entity>
   //
 
   /**
-   * Attached {@link EntityMetadata} captures native details to simplify resolution w/o encoding.
+   * Attach metadata to entity.
    */
-  protected static class AttachedEntityMetadata
-      implements EntityMetadata
-  {
-    private final EntityAdapter owner;
-
-    private ODocument document;
-
-    private EntityId id;
-
-    private EntityVersion version;
-
-    public AttachedEntityMetadata(final EntityAdapter owner, final ODocument document) {
-      this.owner = checkNotNull(owner);
-      this.document = checkNotNull(document);
-      checkNotNull(document);
-    }
-
-    @Override
-    public EntityId getId() {
-      if (id == null) {
-        id = new AttachedEntityId(document, owner);
-      }
-      return id;
-    }
-
-    protected ORecordVersion getRecordVersion() {
-      return document.getRecordVersion();
-    }
-
-    @Override
-    public EntityVersion getVersion() {
-      if (version == null) {
-        version = new EntityVersion(getRecordVersion().toString());
-      }
-      return version;
-    }
-
-    @Override
-    public String toString() {
-      return getClass().getSimpleName() + "{" +
-          "id=" + id +
-          ", version=" + version +
-          '}';
-    }
-  }
-
-  /**
-   * An {@link EntityId} that remains connected to the underlying ODocument. This is necessary with OrientDb
-   * transactions mode, as ODocument ids change when the transaction is committed.
-   */
-  protected static class AttachedEntityId
-      extends EntityId
-  {
-    private final EntityAdapter owner;
-
-    private ODocument document;
-
-    private String cachedValue;
-
-    public AttachedEntityId(final ODocument document, final EntityAdapter owner) {
-      super("");
-      this.owner = checkNotNull(owner);
-      this.document = checkNotNull(document);
-    }
-
-    @Nonnull
-    @Override
-    public String getValue() {
-      if (cachedValue == null) {
-        final ORID identity = document.getIdentity();
-        checkState(!identity.isTemporary(), "attempted use of temporary/uncommitted document id");
-        cachedValue = owner.getRecordIdObfuscator().encode(owner.getType(), identity);
-      }
-      return cachedValue;
-    }
-
-    protected ORID recordIdentity() {
-      return document.getIdentity();
-    }
-  }
-
-  /**
-   * Set metadata on entity.
-   */
-  protected void setMetadata(final T entity, final ODocument doc) {
-    checkNotNull(entity);
+  protected void attachMetadata(final T entity, final ODocument doc) {
     entity.setEntityMetadata(new AttachedEntityMetadata(this, doc));
-  }
-
-  /**
-   * Get metadata from entity.
-   */
-  protected EntityMetadata getMetadata(final T entity) {
-    return metadata(entity);
   }
 
   public EntityId encode(final ORID id) {
@@ -321,7 +224,8 @@ public abstract class EntityAdapter<T extends Entity>
   }
 
   public ORID recordIdentity(final EntityId id) {
-    if (checkNotNull(id) instanceof AttachedEntityId) {
+    checkNotNull(id);
+    if (id instanceof AttachedEntityId) {
       return ((AttachedEntityId) id).recordIdentity();
     }
     return getRecordIdObfuscator().decode(getType(), id.toString());
